Added support for specifying the return value of controller methods in ctx.Trigger.
Previously, the methods specified in trigger would have their return values ignored in the pipeline, which was not suitable for some use cases.
Now, return values are supported. The first parameter of the return value must be of type mvm.Result. By returning a result, you can now exit the pipeline early with the desired result.